requirements.txt
selenium
tencentcloud-sdk-python
opencv-python
pillow